[Backport] 8217362: Emergency dump does not work when disk=false is set
Summary:

Test Plan: jdk/jfr


Issue: https://github.com/alibaba/dragonwell8/issues/112
